Create simple-bulk-inscription-ordinals-for-dummies
This is a Python script that inscribes images on the Bitcoin blockchain using the OpenTimestamps service provided by the Ord tool. The script takes a list of image paths and loops through them, inscribing each one on the Bitcoin blockchain.

The subprocess module is used to execute the ord command-line tool, passing in the appropriate arguments such as the path to the image, the wallet address to use for the inscription, and the transaction fee rate to use.

In this particular script, the image paths are hardcoded in the image_paths list, and the range of images to be inscribed is from 11 to 50. The --cookie-file argument specifies the path to the Bitcoin Core cookie file, which is necessary for connecting to the Bitcoin Core wallet. The --fee-rate argument specifies the transaction fee rate to use for the inscription.

Overall, this script demonstrates a simple way to automate the inscription of images on the Bitcoin blockchain using the Ord tool.